lets print the size of all data types

+++ 0x00-hello_world/6-size.c
@@ -9,7 +9,7 @@ int main(void)
 char c;
 int i;
 long li;
-long lli;
+long  long int lli;
 float f;
 
 printf("Size of a char: %ld byte(s)\n", sizeof(c));
